requirements_new.txt
# Core Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Alternative per siti dinamici
selenium>=4.15.0
playwright>=1.40.0
webdriver-manager>=4.0.0

# Data Management
pandas>=2.1.0
openpyxl>=3.1.0

# Date Parsing
python-dateutil>=2.8.2

# HTTP Headers e User-Agent
fake-useragent>=1.4.0

# Logging migliorato
colorlog>=6.8.0

# Rate Limiting
ratelimit>=2.2.1

# HTML Sanitization
bleach>=6.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# ==============================================================================
# NUOVE LIBRERIE PER IL SISTEMA INTELLIGENTE
# ==============================================================================

# Document Processing - PDF, DOCX, etc.
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
pypandoc>=1.13

# AI Analysis - OpenAI and Anthropic
openai>=1.7.0
anthropic>=0.8.0

# Advanced NLP
spacy>=3.7.0
transformers>=4.36.0

# Topic Modeling
scikit-learn>=1.3.0
gensim>=4.3.0

# Database for Document Storage
sqlalchemy>=2.0.0
alembic>=1.13.0

# File Hashing for Deduplication
hashlib-additional>=1.0.0

# JSON Schema Validation
jsonschema>=4.20.0

# HTML Generation
jinja2>=3.1.0
markdown>=3.5.0

# ChromaDB for Vector Storage (optional for advanced AI)
chromadb>=0.4.0
